Trabajo de la semana 2 del Sprint 3
Terminado el trabajo asignado para este Sprint. Se ha hecho aparte la gestión de veterinarios por Admin y se han empezado los tests funcionales. También se ha empezado un poco los css.

+++ src/main/java/org/springframework/samples/petclinic/user/UserService.java
@@ -27,16 +27,11 @@ import org.springframework.samples.petclinic.owner.Owner;
 import org.springframework.samples.petclinic.owner.OwnerService;
 import org.springframework.samples.petclinic.vet.Vet;
 import org.springframework.samples.petclinic.vet.VetService;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-/**
- * Mostly used as a facade for all Petclinic controllers Also a placeholder
- * for @Transactional and @Cacheable annotations
- *
- * @author Michael Isvy
- */
 @Service
 public class UserService {
 
@@ -71,8 +66,12 @@ public class UserService {
 
 	@Transactional(readOnly = true)
 	public User findCurrentUser() {
-		return userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication().getName())
-				.orElseThrow(() -> new ResourceNotFoundException("User", "Logged In", true));
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		if (auth == null)
+			throw new ResourceNotFoundException("Nobody authenticated!");
+		else
+			return userRepository.findByUsername(auth.getName())
+					.orElseThrow(() -> new ResourceNotFoundException("User", "Username", auth.getName()));
 	}
 
 	public Boolean existsUser(String username) {
@@ -93,27 +92,11 @@ public class UserService {
 		return toUpdate;
 	}
 
-	@Transactional(readOnly = true)
-	public Owner findOwnerByUser(String username) throws DataAccessException {
-		return this.userRepository.findOwnerByUser(username)
-				.orElseThrow(() -> new ResourceNotFoundException("Owner", "User", username));
-	}
-
-//	@Transactional
-//	public void deleteUser(String username) {
-//		try {
-//			deleteRelations(username);		
-//		}catch(ResourceNotFoundException e) {
-//			System.out.println("Owner already deleted. Deleting user.");
-//		}
-//		User toDelete = findUser(username);
-//		userRepository.delete(toDelete);
-//	}
-
 	@Transactional
-	public void deleteUser(Integer id){
+	public void deleteUser(Integer id) {
 		User toDelete = findUser(id);
-		deleteRelations(id, toDelete.getAuthority().getAuthority());
+		if (toDelete.getAuthority() != null)
+			deleteRelations(id, toDelete.getAuthority().getAuthority());
 		userRepository.delete(toDelete);
 	}
 
@@ -121,11 +104,13 @@ public class UserService {
 		switch (auth) {
 		case "OWNER":
 			Optional<Owner> owner = ownerService.optFindOwnerByUser(id);
-			if(owner.isPresent()) ownerService.deleteOwner(owner.get().getId());
+			if (owner.isPresent())
+				ownerService.deleteOwner(owner.get().getId());
 			break;
 		case "VET":
 			Optional<Vet> vet = vetService.optFindVetByUser(id);
-			if(vet.isPresent()) vetService.deleteVet(vet.get().getId());
+			if (vet.isPresent())
+				vetService.deleteVet(vet.get().getId());
 			break;
 		}
